+++ CHANGELOG.md
@@ -47,6 +47,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - [Resources] Functions in the `resources` module no longer have default values for the `version` argument. [#297]
 
+- [Compatibility] Drop support for Python 2.7.
+
 ### Fixed
 
 - [Constants] `STANDARD_VERSIONS` now lists all versions of the Standard, not just those that are fully supported by pyIATI. [#223]